feat: moved ar tests to its own folder
