+++ firnas_vm/tests/ar/mod.rs
@@ -1,0 +1,4 @@
+pub mod class_tests;
+pub mod control_flow_tests;
+pub mod function_tests;
+pub mod simple_tests;
