+++ firnas_vm/tests/integration.rs
@@ -1,5 +1,6 @@
-#![cfg(feature = "en")]
-
 #[macro_use]
 mod common;
+#[cfg(feature = "ar")]
+mod ar;
+#[cfg(feature = "en")]
 mod en;
